+++ src/components/teacher/class-detail/StudentGradingPanel.jsx
@@ -206,6 +206,9 @@ export default function StudentGradingPanel({ submissionPngUrl, referenceKeyBase
           check_breadboard: true,
           check_overlap: true,
           ignore_pin_changes: true
+        },
+        config: {
+          compilerUrl: `${import.meta.env.VITE_API_BASE_URL || 'http://localhost:5001/api'}/compile`
         }
       });
